refactor: reorganize component imports to improve module structure
