+++ src/app/dashboard/layout.tsx
@@ -1,7 +1,7 @@
-import { AppSidebar } from "@/components/sidebar/app-sidebar";
-import Footer from "@/components/footer/Footer";
-import Header from "@/components/header/Header";
+import { AppSidebar } from "@/components/layout/sidebar/app-sidebar";
+import Footer from "@/components/layout/footer/Footer";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
+import Header from "@/components/layout/header/Header";
 
 const layout = ({ children }: { children: React.ReactNode }) => {
   return (
